+++ Editor/InterhapticsMenu.cs
@@ -1,5 +1,5 @@
 ﻿/* ​
-* Copyright © 2022 Go Touch VR SAS. All rights reserved. ​
+* Copyright (c) 2023 Go Touch VR SAS. All rights reserved. ​
 * ​
 */
 
@@ -31,7 +31,8 @@ namespace Interhaptics.Editor
             }
 
             camera.gameObject.AddComponent<HapticManager>();
-            UnityEngine.Debug.Log("An HapticManager was correctly generated");
+            UnityEngine.Debug.Log("Haptic Manager was correctly instantiated.");
+            UnityEditor.SceneManagement.EditorSceneManager.SaveScene(UnityEditor.SceneManagement.EditorSceneManager.GetActiveScene());
         }
 
     }